+++ tracker/npm_package/plausible.d.ts
@@ -44,6 +44,10 @@ export interface PlausibleConfig {
   // This can be used to avoid sending certain types of events, or modifying any event
   // parameters, e.g. to clean URLs of values that should not be recorded.
   transformRequest?: (payload: PlausibleRequestPayload) => PlausibleRequestPayload | null
+
+  // If enabled (the default), the script will set `window.plausible` after `init` is called.
+  // This is used by the verifier to detect if the script is loaded from npm package.
+  bindToWindow?: boolean
 }
 
 export interface PlausibleEventOptions {