+++ tracker/src/plausible.js
@@ -34,21 +34,22 @@ function init(overrides) {
 
     window.plausible = track
     window.plausible.init = init
-    window.plausible.l = true
     window.plausible.v = COMPILE_TRACKER_SCRIPT_VERSION
-
+    
     if (COMPILE_PLAUSIBLE_WEB) {
       window.plausible.s = 'web'
     }
+
+    window.plausible.l = true
   }
 
   // Bind to window to be detectable by the verifier tool
   // This is done in a 'safe' way to avoid breaking the page if window is frozen or running without window
   if (COMPILE_PLAUSIBLE_NPM && config.bindToWindow && typeof window !== 'undefined') {
     window.plausible = track
-    window.plausible.l = true
     window.plausible.s = 'npm'
     window.plausible.v = COMPILE_TRACKER_SCRIPT_VERSION
+    window.plausible.l = true
   }
 }