+++ tracker/npm_package/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.3.4] - 2025-07-23
+
 - Plausible loaded indicator `window.plausible.l = true` is set last in initialisation functions
 
 ## [0.3.3] - 2025-07-22